Finalize the WebviewView Api
Fixes #46585

This new api allows extensions to contribute webviews to the sidebar or panel
